requirements.txt
# Antes de instalar, asegúrate de tener cmake instalado, desde la página oficial:
# https://cmake.org/download/
# O si usas Ubuntu/Debian:
# sudo apt-get install cmake
# O si usas macOS con Homebrew:
# brew install cmake 
# luego verifica si pip esta actualizado:
# pip install --upgrade pip
# Luego instala las dependencias con:
# pip install -r requirements.txt
# Dependencias necesarias para el proyecto de reconocimiento facial (FastAPI, face_recognition, PostgreSQL)

opencv-python==4.13.0.90
Pillow==12.1.0
psycopg2-binary==2.9.11
numpy==2.4.1
tf-keras==2.20.1
deepface==0.0.97
retina-face==0.0.17
pandas==3.0.0
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.6
requests==2.32.5